+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # CHANGELOG
 All notable changes to this project will be documented in this file.
 
+## v1.2.0 (2021-04-02)
+***
+### Added
+- Add a new condition type: `\Zahran\Mapper\Condition\Missing`
+- Added tests cases for Condition types: `Contains` & `Missing`
+
+### Changed
+- Modified `\Zahran\Mapper\Condition\Contains` to allow users to search for:
+  - An occurrence in the source array.
+  - An array of occurrences in the source array.
+  - An occurrence in a source text.
+  - An array of occurrences in the source text.
+
 ## v1.1.0 (2021-04-01)
 ***
 ### Added